Revert (size_t) casts

+++ buddy_alloc.h
@@ -1124,7 +1124,7 @@ static struct buddy_tree *buddy_tree_init(unsigned char *at, uint8_t order) {
     struct buddy_tree *t = (struct buddy_tree*) at;
     memset(at, 0, size);
     t->order = order;
-    t->upper_pos_bound = ((size_t)1u) << t->order;
+    t->upper_pos_bound = 1u << t->order;
     buddy_tree_populate_size_for_order(t);
     return t;
 }
@@ -1154,7 +1154,7 @@ static void buddy_tree_grow(struct buddy_tree *t, uint8_t desired_order) {
                 t->order + 1u, next_pos);
 
             /* There are this many nodes at the current level */
-            size_t node_count = ((size_t)1u) << (current_order - 1u);
+            size_t node_count = 1u << (current_order - 1u);
 
             /* Transfer the bits*/
             bitset_shift_right(buddy_tree_bits(t),
@@ -1174,7 +1174,7 @@ static void buddy_tree_grow(struct buddy_tree *t, uint8_t desired_order) {
         }
         /* Advance the order and refrest the root */
         t->order += 1u;
-        t->upper_pos_bound = ((size_t)1u) << t->order;
+        t->upper_pos_bound = 1u << t->order;
         buddy_tree_populate_size_for_order(t);
 
         /* Update the root */
@@ -1202,7 +1202,7 @@ static void buddy_tree_shrink(struct buddy_tree *t, uint8_t desired_order) {
                 next_order, buddy_tree_parent(left_start));
 
             /* There are this many nodes at the current level */
-            size_t node_count = ((size_t)1u) << (left_start.depth - 1u);
+            size_t node_count = 1u << (left_start.depth - 1u);
 
             /* Transfer the bits*/
             bitset_shift_left(buddy_tree_bits(t),
@@ -1216,7 +1216,7 @@ static void buddy_tree_shrink(struct buddy_tree *t, uint8_t desired_order) {
 
         /* Advance the order */
         t->order = (uint8_t) next_order;
-        t->upper_pos_bound = ((size_t)1u) << t->order;
+        t->upper_pos_bound = 1u << t->order;
         buddy_tree_populate_size_for_order(t);
     }
 }
@@ -1240,7 +1240,7 @@ static struct buddy_tree_pos buddy_tree_leftmost_child(struct buddy_tree *t) {
 static struct buddy_tree_pos buddy_tree_leftmost_child_internal(size_t tree_order) {
     assert(tree_order);
     struct buddy_tree_pos result;
-    result.index = ((size_t)1u) << (tree_order - 1u);
+    result.index = 1u << (tree_order - 1u);
     result.depth = tree_order;
     return result;
 }
@@ -1291,7 +1291,7 @@ static inline size_t buddy_tree_index_internal(struct buddy_tree_pos pos) {
     /* % ((sizeof(size_t) * CHAR_BIT)-1) ensures we don't shift into
      * undefined behavior and stops clang from barking :)
      * Hopefully clang also optimizes it away :) */
-    size_t mask = ((size_t)1u) << (pos.depth - 1) % ((sizeof(size_t) * CHAR_BIT)-1);
+    size_t mask = 1u << (pos.depth - 1) % ((sizeof(size_t) * CHAR_BIT)-1);
     size_t result = pos.index & ~mask;
     return result;
 }
@@ -1596,7 +1596,7 @@ static float buddy_tree_fragmentation(struct buddy_tree *t) {
         size_t pos_status = buddy_tree_status(t, state.current_pos);
         if (pos_status == 0) {
             // Empty node, process
-            size_t virtual_size = ((size_t)1u) << ((tree_order - state.current_pos.depth) % ((sizeof(size_t) * CHAR_BIT)-1));
+            size_t virtual_size = 1u << ((tree_order - state.current_pos.depth) % ((sizeof(size_t) * CHAR_BIT)-1));
             quality += (virtual_size * virtual_size);
             total_free_size += virtual_size;
             // Ascend
